+++ packages/core/styles/src/packages/composite/segmented-control/index.ts
@@ -1,0 +1,9 @@
+import { buttonStyles } from "./nimbus-segmented-control.css";
+
+export const segmentedControl = {
+  subcomponents: {
+    button: {
+      classnames: { ...buttonStyles },
+    },
+  },
+};