+++ packages/react/src/composite/SegmentedControl/src/components/SegmentedControlButton/components/SegmentedControlButtonSkeleton/SegmentedControlButtonSkeleton.types.ts
@@ -1,0 +1,6 @@
+export interface SegmentedControlButtonSkeletonProps {
+  /** Width of the skeleton button */
+  width?: string;
+  /** Height of the skeleton button */
+  height?: string;
+}