+++ packages/react/CHANGELOG.md
@@ -2,15 +2,21 @@
 
 Nimbus is an open-source Design System created by Tiendanube / Nuvesmhop’s team to empower and enhance more stories every day, with simplicity, accessibility, consistency and performance.
 
+## 2025-05-06 `5.11.0`
+
+#### 🎉 New features
+
+- Created `SegmentedControl` new composite component. ([#290](https://github.com/TiendaNube/nimbus-design-system/pull/290) by [@joacotornello](https://github.com/joacotornello))
+
 ## 2025-04-03 `5.10.0`
 
-#### 🐛 New features
+#### 🎉 New features
 
 - `Tabs`: Adds Controlled component support, with the new `selected` and `onTabSelect` props. ([#282](https://github.com/TiendaNube/nimbus-design-system/pull/282) by [@joacotornello](https://github.com/joacotornello))
 
 ## 2025-03-25 `5.9.0`
 
-#### 🐛 New features
+#### 🎉 New features
 
 - Added `textDecoration` property to `Text` component. ([#279](https://github.com/TiendaNube/nimbus-design-system/pull/279) by [@guidobotta](https://github.com/guidobotta))
 - Added `line-through` option to `textDecoration` property. ([#279](https://github.com/TiendaNube/nimbus-design-system/pull/279) by [@guidobotta](https://github.com/guidobotta))